requirements.txt
openai
tqdm
python-dotenv